requirements.txt
smbus2==0.4.3
adafruit-circuitpython-character-lcd==1.3.0
board==1.0
busio==1.0
digitalio==1.0 